+++ app/src/main/java/hu/kts/cmetronome/functional/LongConsumer.java
@@ -3,12 +3,12 @@ package hu.kts.cmetronome.functional;
 /**
  * Created by andrasnemeth on 12/01/16.
  */
-public interface IntConsumer {
+public interface LongConsumer {
 
     /**
      * Performs this operation on the given argument.
      *
      * @param value the input argument
      */
-    void accept(int value);
+    void accept(long value);
 }
